+++ src/main/java/kr/codesqaud/cafe/SessionConst.java
@@ -1,4 +1,4 @@
-package kr.codesqaud.cafe.controller;
+package kr.codesqaud.cafe;
 
 public class SessionConst {
     public static final String LOGIN_USER = "loginUser";
